remove: fmod folder remove: var use add: file helper in utilities class for cross platform loading with documentation
