+++ DewdropEngine/Audio/StreamedSound.cs
@@ -72,13 +72,13 @@ namespace Dewdrop.Audio
         }
 
         /// <summary>
-        /// Sound length milliseconds
+        /// Sound length in milliseconds
         /// </summary>
         public uint Length
         {
             get
             {
-                _nativeSound.getLength(out var length, TIMEUNIT.MS);
+                _nativeSound.getLength(out uint length, TIMEUNIT.MS);
                 return length;
             }
         }
@@ -93,7 +93,7 @@ namespace Dewdrop.Audio
         {
             get
             {
-                _nativeChannel.getLoopCount(out var loops);
+                _nativeChannel.getLoopCount(out int loops);
                 return loops;
             }
             set
@@ -148,7 +148,7 @@ namespace Dewdrop.Audio
         {
             get
             {
-                _nativeChannel.isPlaying(out var isPlaying);
+                _nativeChannel.isPlaying(out bool isPlaying);
                 return isPlaying;
             }
         }
@@ -163,7 +163,7 @@ namespace Dewdrop.Audio
         {
             get
             {
-                _nativeChannel.getPitch(out var pitch);
+                _nativeChannel.getPitch(out float pitch);
                 return pitch;
             }
             set => _nativeChannel.setPitch(value);
@@ -178,7 +178,7 @@ namespace Dewdrop.Audio
         {
             get
             {
-                _nativeChannel.getVolume(out var volume);
+                _nativeChannel.getVolume(out float volume);
                 return volume;
             }
             set => _nativeChannel.setVolume(value);
@@ -193,7 +193,7 @@ namespace Dewdrop.Audio
         {
             get
             {
-                _nativeChannel.getLowPassGain(out var lowPassGain);
+                _nativeChannel.getLowPassGain(out float lowPassGain);
                 return lowPassGain;
             }
             set => _nativeChannel.setLowPassGain(value);
@@ -206,7 +206,7 @@ namespace Dewdrop.Audio
         {
             get
             {
-                _nativeChannel.getMute(out var mute);
+                _nativeChannel.getMute(out bool mute);
                 return mute;
             }
             set => _nativeChannel.setMute(value);
